+++ tiles/src/main/java/com/protomaps/basemap/Basemap.java
@@ -86,7 +86,7 @@ public class Basemap extends ForwardingProfile {
 
   @Override
   public String version() {
-    return "4.0.0-alpha.4";
+    return "4.0.0";
   }
 
   @Override