+++ tiles/src/main/java/com/protomaps/basemap/layers/Roads.java
@@ -8,7 +8,6 @@ import com.onthegomap.planetiler.ForwardingProfile;
 import com.onthegomap.planetiler.VectorTile;
 import com.onthegomap.planetiler.geo.GeometryException;
 import com.onthegomap.planetiler.reader.SourceFeature;
-import com.onthegomap.planetiler.util.Parse;
 import com.protomaps.basemap.feature.FeatureId;
 import com.protomaps.basemap.locales.CartographicLocale;
 import com.protomaps.basemap.locales.US;
@@ -76,7 +75,7 @@ public class Roads implements ForwardingProfile.LayerPostProcesser {
         minZoomNames = 12;
       } else if (highway.equals("secondary") || highway.equals("secondary_link") || highway.equals("tertiary") ||
         highway.equals("tertiary_link")) {
-        kind = "medium_road";
+        kind = "major_road";
         minZoom = 9;
 
         if (highway.equals("secondary")) {
@@ -134,23 +133,20 @@ public class Roads implements ForwardingProfile.LayerPostProcesser {
 
       var feat = features.line("roads")
         .setId(FeatureId.create(sf))
-        // Core Tilezen schema properties
         .setAttr("kind", kind)
         // To power better client label collisions
         .setAttr("min_zoom", minZoom + 1)
         .setAttrWithMinzoom("ref", shield.text(), minZoomShieldText)
         .setAttrWithMinzoom("shield_text_length", shieldTextLength, minZoomShieldText)
         .setAttrWithMinzoom("network", shield.network(), minZoomShieldText)
-        // Core OSM tags for different kinds of places
-        .setAttrWithMinzoom("layer", Parse.parseIntOrNull(sf.getString("layer")), 12)
         .setAttrWithMinzoom("oneway", sf.getString("oneway"), 14)
         // `highway` is a temporary attribute that gets removed in the post-process step
         .setAttr("highway", highway)
+        .setAttr("sort_rank", 400)
         .setMinPixelSize(0)
         .setPixelTolerance(0)
         .setZoomRange(minZoom, maxZoom);
 
-      // Core Tilezen schema properties
       if (!kindDetail.isEmpty()) {
         feat.setAttr("kind_detail", kindDetail);
       } else {
@@ -164,24 +160,117 @@ public class Roads implements ForwardingProfile.LayerPostProcesser {
 
       if (sf.hasTag("highway", "motorway_link", "trunk_link", "primary_link", "secondary_link",
         "tertiary_link")) {
-        feat.setAttr("link", 1);
+        feat.setAttr("is_link", true);
       }
 
       // Set "brunnel" (bridge / tunnel) property where "level" = 1 is a bridge, 0 is ground level, and -1 is a tunnel
       // Because of MapLibre performance and draw order limitations, generally the boolean is sufficent
-      // See also: "layer" for more complicated ±6 layering for more sophisticated graphics libraries
       if (sf.hasTag("bridge") && !sf.hasTag("bridge", "no")) {
-        feat.setAttrWithMinzoom("level", 1, 12);
+        feat.setAttrWithMinzoom("is_bridge", true, 12);
       } else if (sf.hasTag("tunnel") && !sf.hasTag("tunnel", "no")) {
-        feat.setAttrWithMinzoom("level", -1, 12);
-      } else {
-        feat.setAttrWithMinzoom("level", 0, 12);
+        feat.setAttrWithMinzoom("is_tunnel", true, 12);
       }
 
       // Server sort features so client label collisions are pre-sorted
       feat.setSortKey(minZoom);
 
       OsmNames.setOsmNames(feat, sf, minZoomNames);
+    } // end highway=
+
+    // non-highway features
+    // todo: exclude railway stations, levels
+    if (sf.canBeLine() && (sf.hasTag("railway") ||
+      sf.hasTag("aerialway", "cable_car") ||
+      sf.hasTag("man_made", "pier") ||
+      sf.hasTag("route", "ferry") ||
+      sf.hasTag("aeroway", "runway", "taxiway")) &&
+      (!sf.hasTag("building") /* see https://github.com/protomaps/basemaps/issues/249 */) &&
+      (!sf.hasTag("railway", "abandoned", "razed", "demolished", "removed", "construction", "platform", "proposed"))) {
+
+      int minZoom = 11;
+
+      if (sf.hasTag("aeroway", "runway")) {
+        minZoom = 9;
+      } else if (sf.hasTag("aeroway", "taxiway")) {
+        minZoom = 10;
+      } else if (sf.hasTag("service", "yard", "siding", "crossover")) {
+        minZoom = 13;
+      } else if (sf.hasTag("man_made", "pier")) {
+        minZoom = 13;
+      }
+
+      String kind = "other";
+      String kindDetail = "";
+      if (sf.hasTag("aeroway")) {
+        kind = "aeroway";
+        kindDetail = sf.getString("aeroway");
+      } else if (sf.hasTag("railway", "disused", "funicular", "light_rail", "miniature", "monorail", "narrow_gauge",
+        "preserved", "subway", "tram")) {
+        kind = "rail";
+        kindDetail = sf.getString("railway");
+        minZoom = 14;
+
+        if (sf.hasTag("railway", "disused")) {
+          minZoom = 15;
+        }
+      } else if (sf.hasTag("railway")) {
+        kind = "rail";
+        kindDetail = sf.getString("railway");
+
+        if (kindDetail.equals("service")) {
+          minZoom = 13;
+
+          // eg a rail yard
+          if (sf.hasTag("service")) {
+            minZoom = 14;
+          }
+        }
+      } else if (sf.hasTag("ferry")) {
+        kind = "ferry";
+        kindDetail = sf.getString("ferry");
+      } else if (sf.hasTag("man_made", "pier")) {
+        kind = "path";
+        kindDetail = "pier";
+      } else if (sf.hasTag("aerialway")) {
+        kind = "aerialway";
+        kindDetail = sf.getString("aerialway");
+      }
+
+      var feature = features.line(this.name())
+        .setId(FeatureId.create(sf))
+        .setAttr("kind", kind)
+        // Used for client-side label collisions
+        .setAttr("min_zoom", minZoom + 1)
+        .setAttr("network", sf.getString("network"))
+        .setAttr("ref", sf.getString("ref"))
+        .setAttr("route", sf.getString("route"))
+        .setAttr("service", sf.getString("service"))
+        .setAttr("sort_rank", 400)
+        .setZoomRange(minZoom, 15);
+
+      if (!kindDetail.isEmpty()) {
+        feature.setAttr("kind_detail", kindDetail);
+      }
+
+      // Set "brunnel" (bridge / tunnel) property where "level" = 1 is a bridge, 0 is ground level, and -1 is a tunnel
+      // Because of MapLibre performance and draw order limitations, generally the boolean is sufficent
+      // See also: "layer" for more complicated ±6 layering for more sophisticated graphics libraries
+      if (sf.hasTag("bridge") && !sf.hasTag("bridge", "no")) {
+        feature.setAttrWithMinzoom("is_bridge", true, 12);
+      } else if (sf.hasTag("tunnel") && !sf.hasTag("tunnel", "no")) {
+        feature.setAttrWithMinzoom("is_tunnel", true, 12);
+      }
+
+      // Too many small pier lines otherwise
+      if (kindDetail.equals("pier")) {
+        feature.setMinPixelSize(2);
+      }
+
+      // Server sort features so client label collisions are pre-sorted
+      feature.setSortKey(minZoom);
+
+      // TODO: (nvkelso 20230623) This should be variable, but 12 is better than 0 for line merging
+      OsmNames.setOsmNames(feature, sf, 12);
     }
   }
 
@@ -198,9 +287,6 @@ public class Roads implements ForwardingProfile.LayerPostProcesser {
 
     for (var item : items) {
       item.tags().remove("highway");
-      if (!item.tags().containsKey("level")) {
-        item.tags().put("level", 0);
-      }
     }
 
     items = FeatureMerge.mergeLineStrings(items,